--- src/main/java/KUSITMS/WITHUS/domain/mailTemplate/service/MailTemplateService.java
@@ -1,11 +1,0 @@
-package KUSITMS.WITHUS.domain.mailTemplate.service;
-
-import KUSITMS.WITHUS.domain.mailTemplate.dto.MailTemplateRequestDTO;
-import KUSITMS.WITHUS.domain.mailTemplate.dto.MailTemplateResponseDTO;
-
-import java.util.List;
-
-public interface MailTemplateService {
-    List<MailTemplateResponseDTO.Summary> listAll();
-    MailTemplateResponseDTO.Detail create(MailTemplateRequestDTO.Create dto);
-}
