+++ src/main/java/KUSITMS/WITHUS/domain/template/mailTemplate/controller/MailTemplateController.java
@@ -1,8 +1,8 @@
-package KUSITMS.WITHUS.domain.mailTemplate.controller;
+package KUSITMS.WITHUS.domain.template.mailTemplate.controller;
 
-import KUSITMS.WITHUS.domain.mailTemplate.dto.MailTemplateRequestDTO;
-import KUSITMS.WITHUS.domain.mailTemplate.dto.MailTemplateResponseDTO;
-import KUSITMS.WITHUS.domain.mailTemplate.service.MailTemplateService;
+import KUSITMS.WITHUS.domain.template.mailTemplate.dto.MailTemplateRequestDTO;
+import KUSITMS.WITHUS.domain.template.mailTemplate.dto.MailTemplateResponseDTO;
+import KUSITMS.WITHUS.domain.template.mailTemplate.service.MailTemplateService;
 import KUSITMS.WITHUS.global.response.SuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
