+++ src/main/java/KUSITMS/WITHUS/domain/template/mailTemplate/repository/MailTemplateJpaRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.domain.mailTemplate.repository;
+package KUSITMS.WITHUS.domain.template.mailTemplate.repository;
 
-import KUSITMS.WITHUS.domain.mailTemplate.entity.MailTemplate;
+import KUSITMS.WITHUS.domain.template.mailTemplate.entity.MailTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MailTemplateJpaRepository extends JpaRepository<MailTemplate, Long> {
